requirements.txt
# Core
gradio[oauth,mcp]==6.1.0
pandas==2.3.3
numpy
python-dotenv==1.2.1
matplotlib==3.10.8
seaborn==0.13.2
kagglehub==0.3.13
notebook
ipywidgets

# LangChain & embeddings
langchain-community==0.3.2
sentence-transformers==0.3.9
sentencepiece==0.1.98  # pre-built wheel, avoids build errors
langchain_chroma
transformers